feat: create common header for SideDrawer and UpdateFirmwareModal
